Split join code into the respective commands
