+++ microk8s-resources/wrappers/microk8s-leave.wrapper
@@ -10,13 +10,6 @@ export PYTHONNOUSERSITE=false
 
 source $SNAP/actions/common/utils.sh
 
-if echo "$*" | grep -q -- 'help'; then
-  echo "Usage: microk8s leave"
-  echo ""
-  echo "With microk8s leave the node will depart from the cluster it is in."
-  exit 0
-fi
-
 exit_if_stopped
 exit_if_no_permissions
 
@@ -27,10 +20,4 @@ then
   exit 1
 fi
 
-if [ "$#" -ne "0" ]
-then
-  echo "This command will remove the node from the cluster. No arguments are needed."
-  exit 1
-fi
-
-run_with_sudo preserve_env ${SNAP}/usr/bin/python3 ${SNAP}/scripts/cluster/join.py reset
+run_with_sudo preserve_env ${SNAP}/usr/bin/python3 ${SNAP}/scripts/cluster/leave.py $@
