+++ scripts/cluster/join.py
@@ -4,14 +4,13 @@ import random
 import string
 import subprocess
 import os
-import getopt
 import ssl
 import sys
 import time
 import hashlib
 import http
 
-import netifaces
+import click
 import requests
 import socket
 import shutil
@@ -26,7 +25,8 @@ from common.utils import (
     try_initialise_cni_autodetect_for_clustering,
     service,
     mark_no_cert_reissue,
-    unmark_no_cert_reissue,
+    restart_all_services,
+    get_token,
 )
 
 urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
@@ -155,17 +155,6 @@ def get_connection_info(
     except ssl.SSLError as e:
         print("Peer node verification failed with {}.".format(e))
         exit(4)
-
-
-def usage():
-    print("Join a cluster: microk8s join <master>:<port>/<token> [options]")
-    print("")
-    print("Options:")
-    print(
-        "--skip-verify  skip the certificate verification of the node we are"
-        " joining to (default: false)."
-    )
-    print("--force  force the node removal operation" " (default: false).")
 
 
 def set_arg(key, value, file):
@@ -338,7 +327,7 @@ def mark_worker_node():
     lock_file = "{}/var/lock/clustered.lock".format(snapdata_path)
     open(lock_file, "a").close()
     os.chmod(lock_file, 0o700)
-    services = ["etcd", "apiserver-kicker", "kubelite"]
+    services = ["kubelite", "etcd", "apiserver-kicker", "traefik"]
     for s in services:
         service("restart", s)
 
@@ -369,273 +358,6 @@ def store_base_kubelet_args(args_string):
     try_set_file_permissions(args_file)
 
 
-def reset_current_etcd_installation():
-    """
-    Take a node out of a cluster
-    """
-    lock_file = "{}/var/lock/clustered.lock".format(snapdata_path)
-    if not os.path.isfile(lock_file):
-        print("Not in clustering mode.")
-        exit(2)
-
-    os.remove(lock_file)
-    os.remove(ca_cert_file)
-    os.remove(callback_token_file)
-    os.remove(server_cert_file)
-
-    for config_file in ["kubelet", "flanneld", "kube-proxy"]:
-        shutil.copyfile(
-            "{}/default-args/{}".format(snap_path, config_file),
-            "{}/args/{}".format(snapdata_path, config_file),
-        )
-
-    for user in ["proxy", "kubelet"]:
-        config = "{}/credentials/{}.config".format(snapdata_path, user)
-        shutil.copyfile("{}.backup".format(config), config)
-
-    subprocess.check_call("{}/microk8s-stop.wrapper".format(snap_path).split())
-    waits = 10
-    while waits > 0:
-        try:
-            subprocess.check_call("{}/microk8s-start.wrapper".format(snap_path).split())
-            break
-        except subprocess.CalledProcessError:
-            print("Services not ready to start. Waiting...")
-            time.sleep(5)
-            waits -= 1
-
-    unmark_no_cert_reissue()
-
-
-def reset_current_dqlite_installation():
-    """
-    Take a node out of a dqlite cluster
-    """
-    if is_leader_without_successor():
-        print(
-            "This node currently holds the only copy of the Kubernetes "
-            "database so it cannot leave the cluster."
-        )
-        print(
-            "To remove this node you can either first remove all other "
-            "nodes with 'microk8s remove-node' or"
-        )
-        print("form a highly available cluster by adding at least three nodes.")
-        exit(3)
-
-    # We need to:
-    # 1. Stop the apiserver
-    # 2. Send a DELETE request to any member of the dqlite cluster
-    # 3. wipe out the existing installation
-    my_ep, other_ep = get_dqlite_endpoints()
-
-    service("stop", "apiserver")
-    service("stop", "k8s-dqlite")
-    time.sleep(10)
-
-    delete_dqlite_node(my_ep, other_ep)
-
-    print("Generating new cluster certificates.", flush=True)
-    shutil.rmtree(cluster_dir, ignore_errors=True)
-    os.mkdir(cluster_dir)
-    if os.path.isfile("{}/cluster.crt".format(cluster_backup_dir)):
-        # reuse the certificates we had before the cluster formation
-        shutil.copy(
-            "{}/cluster.crt".format(cluster_backup_dir), "{}/cluster.crt".format(cluster_dir)
-        )
-        shutil.copy(
-            "{}/cluster.key".format(cluster_backup_dir), "{}/cluster.key".format(cluster_dir)
-        )
-    else:
-        # This node never joined a cluster. A cluster was formed around it.
-        hostname = socket.gethostname()  # type: str
-        ip = "127.0.0.1"  # type: str
-        shutil.copy(
-            "{}/microk8s-resources/certs/csr-dqlite.conf.template".format(snap_path),
-            "{}/var/tmp/csr-dqlite.conf".format(snapdata_path),
-        )
-        subprocess.check_call(
-            "{}/bin/sed -i s/HOSTNAME/{}/g {}/var/tmp/csr-dqlite.conf".format(
-                snap_path, hostname, snapdata_path
-            ).split(),
-            stdout=subprocess.DEVNULL,
-            stderr=subprocess.DEVNULL,
-        )
-        subprocess.check_call(
-            "{}/bin/sed -i s/HOSTIP/{}/g  {}/var/tmp/csr-dqlite.conf".format(
-                snap_path, ip, snapdata_path
-            ).split(),
-            stdout=subprocess.DEVNULL,
-            stderr=subprocess.DEVNULL,
-        )
-        subprocess.check_call(
-            "{0}/usr/bin/openssl req -x509 -newkey rsa:4096 -sha256 -days 3650 -nodes "
-            "-keyout {1}/var/kubernetes/backend/cluster.key "
-            "-out {1}/var/kubernetes/backend/cluster.crt "
-            "-subj /CN=k8s -config {1}/var/tmp/csr-dqlite.conf -extensions v3_ext".format(
-                snap_path, snapdata_path
-            ).split(),
-            stdout=subprocess.DEVNULL,
-            stderr=subprocess.DEVNULL,
-        )
-
-    # We reset to the default port and address
-    init_data = {"Address": "127.0.0.1:19001"}
-    with open("{}/init.yaml".format(cluster_dir), "w") as f:
-        yaml.dump(init_data, f)
-
-    service("start", "k8s-dqlite")
-    service("start", "apiserver")
-
-    waits = 10  # type: int
-    print("Waiting for node to start.", end=" ", flush=True)
-    time.sleep(10)
-    while waits > 0:
-        try:
-            subprocess.check_call(
-                "{}/microk8s-kubectl.wrapper get service/kubernetes".format(snap_path).split(),
-                stdout=subprocess.DEVNULL,
-                stderr=subprocess.DEVNULL,
-            )
-            subprocess.check_call(
-                "{}/microk8s-kubectl.wrapper apply -f {}/args/cni-network/cni.yaml".format(
-                    snap_path, snapdata_path
-                ).split(),
-                stdout=subprocess.DEVNULL,
-                stderr=subprocess.DEVNULL,
-            )
-            break
-        except subprocess.CalledProcessError:
-            print(".", end=" ", flush=True)
-            time.sleep(5)
-            waits -= 1
-    print(" ")
-    unmark_no_cert_reissue()
-    restart_all_services()
-
-
-def delete_dqlite_node(delete_node, dqlite_ep):
-    if len(delete_node) > 0 and "127.0.0.1" not in delete_node[0]:
-        for ep in dqlite_ep:
-            try:
-                cmd = (
-                    "{snappath}/bin/dqlite -s file://{dbdir}/cluster.yaml -c {dbdir}/cluster.crt "
-                    "-k {dbdir}/cluster.key -f json k8s".format(
-                        snappath=snap_path, dbdir=cluster_dir
-                    ).split()
-                )
-                cmd.append(".remove {}".format(delete_node[0]))
-                subprocess.check_output(cmd)
-                break
-            except Exception as err:
-                print("Contacting node {} failed. Error:".format(ep))
-                print(repr(err))
-                exit(2)
-
-
-def get_dqlite_endpoints():
-    """
-    Return the endpoints the current node has on dqlite and the endpoints of the rest of the nodes.
-
-    :return: two lists with the endpoints
-    """
-    out = subprocess.check_output(
-        "{snappath}/bin/dqlite -s file://{dbdir}/cluster.yaml -c {dbdir}/cluster.crt "
-        "-k {dbdir}/cluster.key -f json k8s .cluster".format(
-            snappath=snap_path, dbdir=cluster_dir
-        ).split()
-    )
-    data = json.loads(out.decode())
-    ep_addresses = []
-    for ep in data:
-        ep_addresses.append(ep["Address"])
-    local_ips = []
-    for interface in netifaces.interfaces():
-        if netifaces.AF_INET not in netifaces.ifaddresses(interface):
-            continue
-        for link in netifaces.ifaddresses(interface)[netifaces.AF_INET]:
-            local_ips.append(link["addr"])
-    my_ep = []
-    other_ep = []
-    for ep in ep_addresses:
-        found = False
-        for ip in local_ips:
-            if "{}:".format(ip) in ep:
-                my_ep.append(ep)
-                found = True
-        if not found:
-            other_ep.append(ep)
-
-    return my_ep, other_ep
-
-
-def is_leader_without_successor():
-    """Checks if the current node is safe to be removed.
-
-    Check if this node acts as a leader to a cluster with more than one nodes where there
-    is no other node to take over the leadership.
-
-    :return: True if this node is the leader without a successor.
-    """
-    out = subprocess.check_output(
-        "{snappath}/bin/dqlite -s file://{dbdir}/cluster.yaml -c {dbdir}/cluster.crt "
-        "-k {dbdir}/cluster.key -f json k8s .cluster".format(
-            snappath=snap_path, dbdir=cluster_dir
-        ).split()
-    )
-    voters = 0
-    data = json.loads(out.decode())
-    ep_addresses = []
-    for ep in data:
-        ep_addresses.append((ep["Address"], ep["Role"]))
-        # Role == 0 means we are voters
-        if ep["Role"] == 0:
-            voters += 1
-
-    local_ips = []
-    for interface in netifaces.interfaces():
-        if netifaces.AF_INET not in netifaces.ifaddresses(interface):
-            continue
-        for link in netifaces.ifaddresses(interface)[netifaces.AF_INET]:
-            local_ips.append(link["addr"])
-
-    is_voter = False
-    for ep in ep_addresses:
-        for ip in local_ips:
-            if "{}:".format(ip) in ep[0]:
-                # ep[1] == ep[Role] == 0 means we are voters
-                if ep[1] == 0:
-                    is_voter = True
-
-    if voters == 1 and is_voter and len(ep_addresses) > 1:
-        # We have one voter in the cluster and the current node is the only voter
-        # and there are other nodes that depend on this node.
-        return True
-    else:
-        return False
-
-
-def remove_kubelet_token(node):
-    """
-    Remove a token for a node in the known tokens
-
-    :param node: the name of the node
-    """
-    file = "{}/credentials/known_tokens.csv".format(snapdata_path)
-    backup_file = "{}.backup".format(file)
-    token = "system:node:{}".format(node)
-    # That is a critical section. We need to protect it.
-    with open(backup_file, "w") as back_fp:
-        with open(file, "r") as fp:
-            for _, line in enumerate(fp):
-                if token in line:
-                    continue
-                back_fp.write("{}".format(line))
-
-    try_set_file_permissions(backup_file)
-    shutil.copyfile(backup_file, file)
-
-
 def replace_admin_token(token):
     """
     Replaces the admin token in the known tokens
@@ -655,119 +377,6 @@ def replace_admin_token(token):
 
     try_set_file_permissions(backup_file)
     shutil.copyfile(backup_file, file)
-
-
-def remove_callback_token(node):
-    """
-    Remove a callback token
-
-    :param node: the node
-    """
-    tmp_file = "{}.tmp".format(callback_tokens_file)
-    if not os.path.isfile(callback_tokens_file):
-        open(callback_tokens_file, "a+")
-        os.chmod(callback_tokens_file, 0o600)
-    with open(tmp_file, "w") as backup_fp:
-        os.chmod(tmp_file, 0o600)
-        with open(callback_tokens_file, "r+") as callback_fp:
-            for _, line in enumerate(callback_fp):
-                parts = line.split()
-                if parts[0] == node:
-                    continue
-                else:
-                    backup_fp.write(line)
-
-    try_set_file_permissions(tmp_file)
-    shutil.move(tmp_file, callback_tokens_file)
-
-
-def remove_node(node):
-    try:
-        # Make sure this node exists
-        subprocess.check_call(
-            "{}/microk8s-kubectl.wrapper get no {}".format(snap_path, node).split(),
-            stdout=subprocess.DEVNULL,
-            stderr=subprocess.DEVNULL,
-        )
-    except subprocess.CalledProcessError:
-        print("Node {} does not exist.".format(node))
-        exit(1)
-
-    remove_kubelet_token(node)
-    remove_callback_token(node)
-    subprocess.check_call(
-        "{}/microk8s-kubectl.wrapper delete no {}".format(snap_path, node).split(),
-        stdout=subprocess.DEVNULL,
-        stderr=subprocess.DEVNULL,
-    )
-
-
-def remove_dqlite_node(node, force=False):
-    try:
-        # Make sure this node exists
-        node_info = subprocess.check_output(
-            "{}/microk8s-kubectl.wrapper get no {} -o json".format(snap_path, node).split()
-        )
-        info = json.loads(node_info.decode())
-        node_address = None
-        for a in info["status"]["addresses"]:
-            if a["type"] == "InternalIP":
-                node_address = a["address"]
-                break
-
-        if not node_address:
-            print("Node {} is not part of the cluster.".format(node))
-            exit(1)
-
-        node_ep = None
-        my_ep, other_ep = get_dqlite_endpoints()
-        for ep in other_ep:
-            if ep.startswith("{}:".format(node_address)):
-                node_ep = ep
-
-        if node_ep and force:
-            delete_dqlite_node([node_ep], my_ep)
-        elif node_ep and not force:
-            print(
-                "Removal failed. Node {} is registered with dqlite. "
-                "Please, run first 'microk8s leave' on the departing node. \n"
-                "If the node is not available anymore and will never attempt to join the cluster "
-                "in the future use the '--force' flag \n"
-                "to unregister the node while removing it.".format(node)
-            )
-            exit(1)
-
-    except subprocess.CalledProcessError:
-        print("Node {} does not exist in Kubernetes.".format(node))
-        if force:
-            print("Attempting to remove {} from dqlite.".format(node))
-            # Make sure we do not have the node in dqlite.
-            # We assume the IP is provided to denote the
-            my_ep, other_ep = get_dqlite_endpoints()
-            for ep in other_ep:
-                if ep.startswith("{}:".format(node)):
-                    print("Removing node entry found in dqlite.")
-                    delete_dqlite_node([ep], my_ep)
-        exit(1)
-
-    remove_node(node)
-
-
-def get_token(name, tokens_file="known_tokens.csv"):
-    """
-    Get token from known_tokens file
-
-    :param name: the name of the node
-    :param tokens_file: the file where the tokens should go
-    :returns: the token or None(if name doesn't exist)
-    """
-    file = "{}/credentials/{}".format(snapdata_path, tokens_file)
-    with open(file) as fp:
-        for line in fp:
-            if name in line:
-                parts = line.split(",")
-                return parts[0].rstrip()
-    return None
 
 
 def store_cert(filename, payload):
@@ -845,36 +454,6 @@ def store_callback_token(token):
     with open(callback_token_file, "w") as fp:
         fp.write(token)
     try_set_file_permissions(callback_token_file)
-
-
-def restart_all_services():
-    """
-    Restart all services
-    """
-    waits = 10
-    while waits > 0:
-        try:
-            subprocess.check_call(
-                "{}/microk8s-stop.wrapper".format(snap_path).split(),
-                stdout=subprocess.DEVNULL,
-                stderr=subprocess.DEVNULL,
-            )
-            break
-        except subprocess.CalledProcessError:
-            time.sleep(5)
-            waits -= 1
-    waits = 10
-    while waits > 0:
-        try:
-            subprocess.check_call(
-                "{}/microk8s-start.wrapper".format(snap_path).split(),
-                stdout=subprocess.DEVNULL,
-                stderr=subprocess.DEVNULL,
-            )
-            break
-        except subprocess.CalledProcessError:
-            time.sleep(5)
-            waits -= 1
 
 
 def update_dqlite(cluster_cert, cluster_key, voters, host):
@@ -975,7 +554,8 @@ def join_dqlite(connection_parts, verify=False, worker=False):
 def update_traefik(master_ip, api_port):
     lock_path = os.path.expandvars("${SNAP_DATA}/var/lock")
     lock = "{}/no-traefik".format(lock_path)
-    os.remove(lock)
+    if os.path.exists(lock):
+        os.remove(lock)
 
     traefik_providers = os.path.expandvars("${SNAP_DATA}/args/traefik/provider-template.yaml")
     traefik_providers_out = os.path.expandvars("${SNAP_DATA}/args/traefik/provider.yaml")
@@ -989,23 +569,32 @@ def update_traefik(master_ip, api_port):
     service("restart", "traefik")
 
 
+def print_traefik_usage():
+    """
+    Print Traefik usage
+    """
+    print("")
+    print("The node has joined the cluster and will appear in the nodes list in a few seconds.")
+    print("")
+    print("If you run a multi-master cluster without a VIP please add all the Kubernetes API server endpoints in")
+    print("/var/snap/microk8s/current/args/traefik/provider.yaml and restart the node.")
+    print("")
+
+
 def join_dqlite_worker_node(info, master_ip, token):
     hostname_override = info["hostname_override"]
-    store_cert("ca.crt", info["ca"])
     store_remote_ca(info["ca"])
-    store_cert("ca.key", info["ca_key"])
     store_cert("serviceaccount.key", info["service_account_key"])
 
     store_base_kubelet_args(info["kubelet_args"])
     update_kubeproxy(info["kubeproxy"], info["ca"], "127.0.0.1", info["apiport"], hostname_override)
     update_kubelet(info["kubelet"], info["ca"], "127.0.0.1", info["apiport"])
-    if "admin_token" in info:
-        replace_admin_token(info["admin_token"])
     create_admin_kubeconfig(info["ca"], info["admin_token"])
     store_callback_token(info["callback_token"])
     update_traefik(master_ip, info["apiport"])
     mark_worker_node()
     mark_no_cert_reissue()
+    print_traefik_usage()
 
 
 def join_dqlite_master_node(info, master_ip, token):
@@ -1065,54 +654,28 @@ def join_etcd(connection_parts, verify=True):
     mark_no_cert_reissue()
 
 
-if __name__ == "__main__":
-    try:
-        opts, args = getopt.gnu_getopt(
-            sys.argv[1:], "hfsw", ["help", "force", "skip-verify", "worker"]
-        )
-    except getopt.GetoptError as err:
-        print(err)  # will print something like "option -a not recognized"
-        usage()
-        sys.exit(2)
+@click.command(context_settings={"ignore_unknown_options": True})
+@click.argument('connection', required=True)
+@click.option('--worker', 'worker', default=False, flag_value='as-worker',
+              help='Join as a worker only node.')
+@click.option('--controlplane', 'worker', flag_value='as-master',
+              help='Join running the control plane on HA clusters. (default)')
+@click.option("--skip-verify", is_flag=True, required=False, default=False,
+              help='Skip the certificate verification of the node we are joining to. (default: false)')
+def join(connection, worker, skip_verify):
+    """
+    Join the node to a cluster
 
-    force = False
-    worker = False
-    verify = True
-    for o, a in opts:
-        if o in ("-h", "--help"):
-            usage()
-            sys.exit(1)
-        elif o in ("-f", "--force"):
-            force = True
-        elif o in ("-s", "--skip-verify"):
-            verify = False
-        elif o in ("-w", "--worker"):
-            worker = True
-        else:
-            print("Unhandled option")
-            sys.exit(1)
-
-    if len(args) <= 0:
-        print("Please provide a connection string.")
-        usage()
-        sys.exit(4)
-    elif args[0] == "reset":
-        if len(args) > 1:
-            if is_node_running_dqlite():
-                remove_dqlite_node(args[1], force)
-            else:
-                remove_node(args[1])
-
-        else:
-            if is_node_running_dqlite():
-                reset_current_dqlite_installation()
-            else:
-                reset_current_etcd_installation()
+    CONNECTION: the cluster connection endpoint in format <master>:<port>/<token>
+    """
+    connection_parts = connection.split("/")
+    verify = not skip_verify
+    if is_node_running_dqlite():
+        join_dqlite(connection_parts, verify, worker)
     else:
-        connection_parts = args[0].split("/")
-        if is_node_running_dqlite():
-            join_dqlite(connection_parts, verify, worker)
-        else:
-            join_etcd(connection_parts, verify)
-
+        join_etcd(connection_parts, verify)
     sys.exit(0)
+
+
+if __name__ == "__main__":
+    join(prog_name="microk8s join")
