+++ snap/hooks/configure
@@ -578,7 +578,8 @@ then
 fi
 
 if [ -e "${SNAP_DATA}/args/cni-network/cni.yaml" ] &&
-   [ -e "${SNAP_DATA}/var/lock/ha-cluster" ]
+   [ -e "${SNAP_DATA}/var/lock/ha-cluster" ] &&
+   ! [ -e "${SNAP_DATA}/var/lock/clustered.lock" ]
 then
   echo "Setting up the CNI"
   start_timer="$(date +%s)"
